do not overwrite certs if they exist

+++ AppController/lib/nginx.rb
@@ -715,8 +715,12 @@ CONFIG
 
     # copy over certs for ssl
     # just copy files once to keep certificate as static.
-    HelperFunctions.shell("cp /etc/appscale/certs/mykey.pem #{NGINX_PATH}")
-    HelperFunctions.shell("cp /etc/appscale/certs/mycert.pem #{NGINX_PATH}")
+    if !File.exists?("#{NGINX_PATH}/mykey.pem")
+      HelperFunctions.shell("cp /etc/appscale/certs/mykey.pem #{NGINX_PATH}")
+    end
+    if !File.exists?("#{NGINX_PATH}/mycert.pem")
+      HelperFunctions.shell("cp /etc/appscale/certs/mycert.pem #{NGINX_PATH}")
+    end
     # Write the main configuration file which sets default configuration parameters
     File.open(MAIN_CONFIG_FILE, "w+") { |dest_file| dest_file.write(config) }
   end
